[Aftab2] Add ; Added the commerical data

+++ Oops/Stock.cs
@@ -8,13 +8,14 @@ namespace Oops
 {
     public  class Stock
     {
-        public int Price { get; set; }
-        public int TotalPrice { get; set; }
+        public double Price { get; set; }
+        public double TotalPrice { get; set; }
         public int nostocks { get; set; }
         public string Name { get; set; }
         public string Symbol { get; set; }
 
        
     }
+
     
 }
